Clarify Status vs Spec fields
Previous implementation was unnecessarily mirroring the observed
Privileges and connectionLimit settings into Status - where these
basically represent what the user inputs as long as the queries return
no error.

This code now refers to the Spec fields for those items that are set by
the user, and we now only pass a calculated list of applied privileges
to be stored in the Status field, which is used for display purposes.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>