Remove remaining table priv code
Also add withAdminOption and withGrantOption to add these modifiers to
the relevant permissions.

Add basic tests

Caveat: Grant code as it currently is can only support ONE instance for each
DB / Role pair. Multiple instances will step on each others' toes due to
the way the permissions are detected. This also means that you can't set
e.g. withAdminOption on one permission and also grant others.

It may be necessary to detect grant equivalence locally and implement
the Update method to get around this, but it will add a good chunk of
complexity.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>