Restrict mysql grant privilege strings to pattern
Privileges for the MySQL Grant type are simply a list of unfiltered strings.
Without listing each of these, the simplest way to avoid arbitrary SQL
injection is to restrict the character set of each grant. Grant
identifiers must be uppercase, and contain no special characters other
than underscores and spaces.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>