Anchor grant pattern correctly
Upon further testing with what should be invalid Postgres grant types,
it was obvious that the validation pattern is not automatically
anchored, as it would still allow arbitrary characters to be entered.
Anchoring the pattern at either end forces each privilege to only
consist of the characters specified in the pattern.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>