Fix role privilege reconcile for non superusers
Previously, if we detected a privilege change in a Role resource, we
would generate a list of all of the privileges and apply them. This
included privileges that had not actually changed.
A side-effect of this was that fields that were late-inited would
then trigger an ALTER ROLE query containing all privilege settings -
even settings that they already had as defaults.
It appears that PostgreSQL has slightly different permissions
checks for CREATE ROLE vs ALTER ROLE, which would cause errors when
trying to set NOSUPERUSER on a role that already had NOSUPERUSER from
creation - but only if the provider-sql user was not a superuser (as is
the case on AWS RDS instances and possibly others).

We fix this by comparing the requested list of privileges with the
observed list of privileges and make sure to only run a query to update
the privileges that have changed (if any).

Fixes #25

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>