Fix panic on changedPrivs
If status.AtProvider.PrivilegesAsClauses were edited, or e.g. restored
from a backup, it may be possible to have fewer observed privileges than
desired privileges, which would cause a panic on update. We now check
that the observed privilege list is at least as long as the desired
list, so we can check for changed privileges safely.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>