Fix ResolveReferences on Grant types
It appears that Resolving references from Selectors has not worked
previously for either the MySQL or PostgreSQL Grant types. This is down
to missing references to the `List` for each referenced type, and
further, not setting the `Ref` field to the resolved reference.

This commit adds the missing code taking reference from a known-working
Selector implementation in provider-aws.

Fixes #29

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>