Allow db name to be passed to postgres client
Allow the database name to be overridden on instantiation of the
postgresql client.

This is in preparation for further functionality where the connection
database name for the PostgreSQL client needs to be controlled
dynamically, where resources must be created under a particular
database. We must _connect_ on a particular database to be able to
create resources on that db (e.g. Extensions), so we must be able to
pass a database name into the client.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>