Fix unable to delete extension on deleted db
If an Extension was created on a database reference that was then
deleted, and a user tried to then delete the extension, it would sit in
an eternal observation error loop. Observe now checks for an 'invalid
catalog error' as returned from lib/pq to detect this, and reports a
nonexistent resource if the database itself has been deleted.

Signed-off-by: Ben Agricola <bagricola@squiz.co.uk>