postgres: Expand grant privileges where required (#144)
* postgres: Expand grant privileges where required

Inside postgresql, some privileges are altered[0] when issuing a `GRANT`
- for example `ALL` is expanded to `CREATE, TEMPORARY, CONNECT`.

When we observe the grant, we query privileges on the database to see if
the ones in the grant exist. Given that they are expanded, we can't
simply query for what the user gave us. Expand the privileges which need
expanding before we make the query.

Closes: #92

[0]: https://www.postgresql.org/docs/15/ddl-priv.html

Signed-off-by: Iain Lane <iain.lane@grafana.com>

* Sort the grants privileges before comparing them

When we use them in the observation the order doesn't matter, but in the
test we are generating a diff where it does, so compare them in the
test.

Signed-off-by: Iain Lane <iain@orangesquash.org.uk>

---------

Signed-off-by: Iain Lane <iain.lane@grafana.com>
Signed-off-by: Iain Lane <iain@orangesquash.org.uk>
Signed-off-by: Timotej Avsec <avsectimotej1@gmail.com>