feat(postgresql): add TLS related options
This adds `sslCert`, `sslKey` and `sslRootCert` options to the
PostgreSQL `ProviderConfig` needed to establish a connection to a
database that either uses non-public certificates or requires client
certificate authentication.

I've opted to create the `Options` struct to hold these and the existing
`sslMode` parameter, this will allow adding other options[1] easier in
the future.

[1] https://www.postgresql.org/docs/current/libpq-connect.html#LIBPQ-PARAMKEYWORDS

Signed-off-by: Zoran Regvart <zoran@regvart.com>