feat(postgresql): TLS options via Secret
Adds support for configuring TLS options using the credentials Secret.
Very similar to how other providers are configured, this now also reads
the `clusterCA`, `clientCert` and `clientKey` keys from the Secret.

Integration tests are added, though they could be optimized to setup the
PostgreSQL database only once and modify the configuration as needed to
reduce the lenght of the tests, left as is for now.

Signed-off-by: Zoran Regvart <zoran@regvart.com>